Fixed Table Errors

+++ src/components/TableTodo.tsx
@@ -1,5 +1,5 @@
-import { useState, useRef, useEffect, useCallback } from "react";
-import { Button, Popconfirm, Space, Switch, Table, Tag } from "antd";
+import { useState, useRef, useMemo } from "react";
+import { Button, Popconfirm, Space, Switch, Table } from "antd";
 import type { ColumnsType } from "antd/es/table";
 import ModalForm, { ModalHandles } from "./modal-form/ModalForm";
 import { useTranslation } from "react-i18next";
@@ -8,34 +8,19 @@ import { useFetchSWR } from "../hooks/useFetchSWR";
 import api from "../services/api";
 import dayjs from "dayjs";
 import { DeleteOutlined, EditOutlined } from "@ant-design/icons";
+import { SorterResult } from "antd/es/table/interface";
 
 function TableTodo() {
-  const [editData, setEditData] = useState<TodoType | null>();
   const { t } = useTranslation();
+  const [editData, setEditData] = useState<TodoType | null>();
 
   const [order, setOrder] = useState<string>("desc");
   const [orderColumn, setOrderColumn] = useState<string>("create_at");
   const [currentPage, setCurrentPage] = useState<number>(1);
-  const { data, isLoading, mutate } = useFetchSWR(
+
+  const { data, mutate, isLoading } = useFetchSWR(
     `todos/?_sort=${orderColumn}&_order=${order}&_page=${currentPage}`
   );
-  const [dataTodo, setDataTodo] = useState<TodoType[]>([]);
-  const [totalPages, setTotalPages] = useState<number>();
-
-  const orders = useCallback(
-    (local: string, order: any) => {
-      setOrderColumn(local);
-      setOrder(order === "descend" ? "desc" : "asc");
-    },
-    [order, orderColumn]
-  );
-
-  useEffect(() => {
-    if (data) {
-      setDataTodo(data.data);
-      setTotalPages(+data.headers.get("x-total-count"));
-    }
-  }, [totalPages, data, dataTodo]);
 
   const modalRef = useRef<ModalHandles>(null);
 
@@ -44,16 +29,16 @@ function TableTodo() {
     modalRef.current?.openMyModal();
   }
 
-  async function handleDelete(record: TodoType) {
-    try {
-      api.delete(`todos/${record.id}`);
-      mutate();
-    } catch (error) {
-      console.log(error);
+  async function handleAction(record: TodoType, typeAction: string) {
+    if (typeAction === "delete") {
+      try {
+        api.delete(`todos/${record.id}`);
+        mutate();
+      } catch (error) {
+        console.log(error);
+      }
+      return;
     }
-  }
-
-  async function handleCompleted(record: TodoType) {
     try {
       api.patch(`todos/${record.id}`, {
         completed: true,
@@ -64,61 +49,58 @@ function TableTodo() {
     }
   }
 
-  const columns: ColumnsType<TodoType> = [
-    {
-      title: `${t("todo")}`,
-      dataIndex: "todo",
-      key: "todo",
-      sorter: (a, b, c) => {
-        orders("todo", c);
-        return 0;
+  const columns = useMemo<ColumnsType<TodoType>>(
+    () => [
+      {
+        title: `${t("todo")}`,
+        dataIndex: "todo",
+        key: "todo",
+        sorter: true,
+        render: (text, record) => {
+          return !record.completed ? <p>{text}</p> : <s>{text}</s>;
+        },
       },
-      render: (text, record) => {
-        return !record.completed ? <p>{text}</p> : <s>{text}</s>;
+      {
+        title: `${t("dateComplet")}`,
+        dataIndex: "dateDoTodo",
+        key: "dateDoTodo",
+        sorter: true,
+        render: (text) => {
+          const dateFormat = dayjs(+text).format("DD/MM/YYYY HH:mm:ss");
+          return <p>{dateFormat}</p>;
+        },
       },
-    },
-    {
-      title: `${t("dateComplet")}`,
-      dataIndex: "dateDoTodo",
-      key: "dateDoTodo",
-      sorter: (a, b, c) => {
-        orders("dateDoTodo", c);
-        return 0;
-      },
-      render: (text) => {
-        const dateFormat = dayjs(+text).format("DD/MM/YYYY HH:mm:ss");
-        return <p>{dateFormat}</p>;
-      },
-    },
-    {
-      title: `${t("actionTable")}`,
-      key: "action",
-      render: (_, record) => (
-        <Space size="middle">
-          <Popconfirm
-            title={t("deleteConfirm")}
-            onConfirm={() => handleDelete(record)}
-          >
-            <Button danger icon={<DeleteOutlined />}>
-              {t("delete")}
+      {
+        title: `${t("actionTable")}`,
+        key: "action",
+        render: (_, record) => (
+          <Space size="middle">
+            <Popconfirm
+              title={t("deleteConfirm")}
+              onConfirm={() => handleAction(record, "delete")}
+            >
+              <Button danger icon={<DeleteOutlined />}>
+                {t("delete")}
+              </Button>
+            </Popconfirm>
+            <Button
+              onClick={() => handleEditRow(record)}
+              icon={<EditOutlined />}
+              disabled={record.completed}
+            >
+              {t("editTodo")}
             </Button>
-          </Popconfirm>
-          <Button
-            onClick={() => handleEditRow(record)}
-            icon={<EditOutlined />}
-            disabled={record.completed}
-          >
-            {t("editTodo")}
-          </Button>
-          <Switch
-            disabled={record.completed}
-            checked={record.completed}
-            onClick={() => handleCompleted(record)}
-          />
-        </Space>
-      ),
-    },
-  ];
+            <Switch
+              disabled={record.completed}
+              checked={record.completed}
+              onClick={() => handleAction(record, "complete")}
+            />
+          </Space>
+        ),
+      },
+    ],
+    []
+  );
 
   return (
     <div style={{ display: "flex", justifyContent: "center" }}>
@@ -126,16 +108,19 @@ function TableTodo() {
         style={{ width: "700px" }}
         bordered
         columns={columns}
-        tableLayout="auto"
-        dataSource={dataTodo}
+        dataSource={data ? data.data : []}
         loading={isLoading}
-        onChange={(record) => {
-          console.log(record);
+        onChange={(pagination, filters, sorter) => {
+          const sort = sorter as SorterResult<TodoType>;
+          console.log({ pagination, filters, sort });
+
+          setOrderColumn(sort.field as string);
+          setOrder((sort.order as string) === "ascend" ? "asc" : "desc");
         }}
-        // rowClassName={(record) => (record.completed ? "disabled-row" : "")}
+        rowKey="id"
         pagination={{
           pageSize: 10,
-          total: totalPages,
+          total: data ? +data.headers.get("x-total-count") : 1,
           onChange(page) {
             setCurrentPage(page);
           },
